requirements.txt
requests~=2.32.3
ffmpeg-python~=0.2.0
Flask~=3.1.0
gTTS~=2.5.4